Update main.go
Updated main.go to add more functionality. The search function takes in a query parameter that will be sent by the front end and populates a list of songs that would show up if you search for the name of this song on the Spotify app. This will allow for better user experiences as they can search for songs and pick from a list that will come up.